requirements.txt
# Client
dash>=2.7.0
dash-bootstrap-components>=1.2.1

# API
Flask>=2.2.2
bcrypt>=4.0.1
pymongo>=4.3.2
python-dotenv>=0.21.0
requests>=2.28.1
